chore(nat): Rename "stateless" submodule to "static_nat"
We want to rename the NAT modes: start by changing from the "stateless
NAT" module to "static NAT" instead. No code change in this commit, we
only rename the module and fix the includes.

For renaming, we use "static_nat" rather than just "static", because the
latter is a reserved keyword.
