+++ nat/src/stateful/test.rs
@@ -47,7 +47,7 @@ use tracing::debug;
 use tracing_test::traced_test;
 
 const ONE_MINUTE: Duration = Duration::from_mins(1);
-use crate::stateless::test::build_gwconfig_from_overlay;
+use crate::static_nat::test::build_gwconfig_from_overlay;
 
 fn test_case(msg: &str) {
     debug!("{}", Frame(msg));
